+++ spotinst/ocean_aws_auto_scaling/fields_spotinst_ocean_aws_auto_scaling.go
@@ -175,6 +175,10 @@ func expandAutoscaler(data interface{}, nullify bool) (*aws.AutoScaler, error) {
 		autoscaler.SetCooldown(spotinst.Int(v))
 	}
 
+	if v, ok := m[string(AutoHeadroomPercentage)].(int); ok && v > 0 {
+		autoscaler.SetAutoHeadroomPercentage(spotinst.Int(v))
+	}
+
 	if v, ok := m[string(AutoscaleDown)]; ok {
 		down, err := expandOceanAWSAutoScalerDown(v)
 		if err != nil {